+++ Modules/Intent.Modules.AzureFunctions.Dispatch.MediatR/FactoryExtensions/AzureFunctionsMediatRDispatchInstaller.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Intent.Application.MediatR.Api;
 using Intent.AzureFunctions.Api;
 using Intent.Engine;
 using Intent.Modelers.Services.Api;