+++ Modules/Intent.Modules.AzureFunctions/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 5.0.13
 
 - Fixed: Corrected Exception Logging.
+- Fixed: Nuget package management update.
 
 ### Version 5.0.12